+++ src/AspNet.Security.OAuth.Alipay/AlipayAuthenticationDefaults.cs
@@ -1,4 +1,10 @@
-﻿using Microsoft.AspNetCore.Authentication;
+﻿/* 
+ * Licensed under the Apache License, Version 2.0 (http://www.apache.org/licenses/LICENSE-2.0) 
+ * See https://github.com/aspnet-contrib/AspNet.Security.OAuth.Providers 
+ * for more information concerning the license and the contributors participating to this project. 
+ */
+
+using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Authentication.OAuth;
 
 namespace AspNet.Security.OAuth.Alipay
