+++ src/AspNet.Security.OAuth.Alipay/AlipayAuthenticationExtensions.cs
@@ -1,4 +1,10 @@
-﻿using System;
+﻿/* 
+ * Licensed under the Apache License, Version 2.0 (http://www.apache.org/licenses/LICENSE-2.0) 
+ * See https://github.com/aspnet-contrib/AspNet.Security.OAuth.Providers 
+ * for more information concerning the license and the contributors participating to this project. 
+ */
+
+using System;
 using AspNet.Security.OAuth.Alipay;
 using JetBrains.Annotations;
 using Microsoft.AspNetCore.Authentication;
