+++ src/AspNet.Security.OAuth.Alipay/AspNet.Security.OAuth.Alipay.csproj
@@ -6,24 +6,13 @@
 
   <PropertyGroup>
     <Description>ASP.NET Core security middleware enabling Alipay authentication.</Description>
-    <Authors>vicente</Authors>
+    <Authors>Vicente Yu</Authors>
     <PackageTags>aspnetcore;authentication;oauth;alipay;security</PackageTags>
   </PropertyGroup>
 
   <ItemGroup>
-    <Compile Remove="bin\**" />
-    <Compile Remove="obj\**" />
-    <EmbeddedResource Remove="bin\**" />
-    <EmbeddedResource Remove="obj\**" />
-    <None Remove="bin\**" />
-    <None Remove="obj\**" />
-  </ItemGroup>
-
-  <ItemGroup>
     <FrameworkReference Include="Microsoft.AspNetCore.App" />
+    <PackageReference Include="JetBrains.Annotations" Version="$(JetBrainsVersion)" PrivateAssets="All" />
   </ItemGroup>
 
-  <ItemGroup>
-    <PackageReference Include="JetBrains.Annotations" Version="2019.1.3" />
-  </ItemGroup>
 </Project>
